Restructured the function to avoid a problem where update statistics was not getting run on the new tables.  This was the situation with the old routine:
1. The code at the front of the routine, that populates the tables, would
   run.
2. The transaction that deleted the old tables and renamed the new would
   start and finish.
3. The transaction that updated statistics on the new tables would start.
4. If the request was submitted through the web, it would timeout after
   5 minutes, while in this transaction.
5. The update statistics transaction would roll back, leaving all of the
   new tables without statistics, and the web site would slow to a crawl.

The new structure of the function creates indexes and runs update statistics
as soon as the tables are fully populated.  The update statistics transaction
at the end has also been dropped.  This has 2 impacts:

1. The function runs faster.  Later code in the function benefits from the
   update statistics on tables that are created early in the function.
2. The deleteing and renaming transaction at the end now runs much faster.
   This used to create indexes and now it no longer does.

However, the cost for all this is added complexity in the index creation
code.  Unlike tables, you can't rename indexes. Each table now has 2 sets
of index creation code, and which one gets executed depends on which set
of names is already in use from the last run.


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@2589 908b024a-c82b-635c-c293-f2b15b23277c