In revision 1.61 of this file, in addition to what I described when I checked the file in, I also converted several "conc" calls to use the SQL concatenate operator, "||", instead.
However, it turns out that "conc" and "||" are not the same.  "conc" deals
with NULLs by treating them as if they were the empty string.  "||" deals
with NULLs by converting the whole result to a NULL string.

Fixed the offending code to do the right thing when encountering NULLs, and
still use the SQL concatenate operator.


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@2720 908b024a-c82b-635c-c293-f2b15b23277c