got around a bug where thru the perl DBI
let x = hi = lo

(location >= lo) AND (location <= hi) is not equivilant to (location == x)

which would be the  mathematicaly correct thing to do.

So in the case where hi == lo I added a tiny amount (.001) to hi and subtracted
the tiny amount to lo and the query works correctly.


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@3249 908b024a-c82b-635c-c293-f2b15b23277c