BUGZID: 123 Dropped code that set the ownerColumn and zdbIdColumn variables when calling other app pages.  These were eventually used in secure_navigation.apg to deal with tables (such as fish_image) that had owner and ZDB ID columns that had names other than "owner" and "zdb_id".
However, we have modified secure_navigation.apg to set the owner and
ZDB ID column names locally, based on the table name passed in.  This is
much simpler.

Also dropped code that set the owner and ZDB ID column names in the
return_recs table.  It was used in do-select.apg for similar purposes.
do-select.apg now sets the column names locally as well.


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@3322 908b024a-c82b-635c-c293-f2b15b23277c