BUGZID: 273 Added comment that this page is no longer in use. Not sure for how long or why.
Dropped insert statement in front of drop table statement.  This hack no
longer works in informix 9.4.  Had to set the MIqry2pass variable in the
webdriver.  Added a comment that this page cannot be webexploded from other
pages.  In the long term, we should convert the temp table to a
permanent table, and then we could drop the MIqry2pass variable.


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@3622 908b024a-c82b-635c-c293-f2b15b23277c