BUGZID: 273 Converted the page from using temporary tables to using app page specific permanent tables that emulate temporary tables by storing the session id with every record.
This was necessary because we were exploiting a hack in Informix that went
away in 9.4.  Informix 9.3 allowed us to drop temp tables in app pages that
were being web exploded.  See the discussion of the MIqry2pass variable in
the web datablade documentation for an explanation of what is going on.


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@3623 908b024a-c82b-635c-c293-f2b15b23277c