BUGZID: 454 For years this script has existed on chromix as /private/bin/domainlookup.pl. Its purpose was/is to replace IP addresses in apache logs with the domain names that correspond to them.  We don't log domain names in the first place because it is to expensive.  It is cheaper just to do the lookup once a month.
As part of the work to replace chromix with a new machine, I am moving this
and other scripts from the chromix filesystem to be under CVS control.

The first revision of this file is a barely modified copy of the file from
chromix.  All I changed were comments and the path to perl (to make the
path consistent with other scripts).


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@4454 908b024a-c82b-635c-c293-f2b15b23277c