BUGZID: none All I set out to do on this revision was to change the description of when GO data is updated.  Then, just for kicks, I validated the page.  It turned out to have lots of problems, so I fixed them.  Then an odd thing happened. The text at the top of the page, describing the page, was now exteneding out of the browser window to the width of the widest table on the page.
With Kevin's help, we figured out what was going on.

The last thing header.js does is open a table, row, and column.  The old
downloads.html appears to have accidentally closed the table at line 172.
Then at line 212, it closes a non-existant table.

The first thing that footer.js does is close a table.  In the old
downloads.html, there was no table to close.

This revision closes the table at the top of the page, and then opens a div
with equivalent properties.  At the bottom of the page, it closes the div, and
opens the table.  In an ironic twist, the closing and opening of the
tables causes the page to no longer validate.  This change was marked with
a :KLUDGE: tag.

An alternative to this approach would have been to modify header.js and
footer.js to open and close a div instead of a table.  This is a good long
term plan, but it was rejected as too big just to get this fix in.


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@4656 908b024a-c82b-635c-c293-f2b15b23277c