BUGZID: none Reviewed by: Tom Updated temp table columns for mrkr_abbrev to use its new length of varchar(40).
In the process of doing this I split the gene column in the manual_genpept
table into two columns: gene and gene_abbrev.  Previously the gene column
first contained a gene abbrev, and was then replaced by a gene ZDB ID.

This worked, but it was a little confusing, and it was dangerous in the long
run.  The gene column was defined as varchar(50), which is currently big
enough to hold ZDB IDs (50) or gene abbrevs (40).  However, if gene abbrevs
grow to 60 in the future then we would probably miss this file.


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@4745 908b024a-c82b-635c-c293-f2b15b23277c