BUGZID: 461 Reviewed by: Peiran This function is the successor to the regen_genomics function.  regen_genomics used to create a wide variety of fast search tables, but lately it has only created all_map_names.
This function was created when the all_name_ends table and incremental regens
were added.

This function recreates the entire all_map_names and all_name_ends table.
This is in contrast to the regen_names_{fish,locus,marker} functions which
replace only the names for a given ZDB ID.

The initial version of this file has a dramatic decline in ease of debugging.
There are two things causing this:
1. Most of the work is now done by subroutines and I can't figure out how
   to get the errorHint mechanism to work inside a subroutine.
2. The work tables are now temporary tables instead of permanent.  You can't
   look at their contents after an abort to help you figure out where things
   went awry.  The work tables have to be temporary so that the 4 regen_names
   routines don't collide with each other.


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@4920 908b024a-c82b-635c-c293-f2b15b23277c