BUGZID: 461 Reviewed by: Peiran xfig drawing showing the call structure of the regen_names* functions, and what routines access which tables.
This figure uses xfig "depths".  Each different type of data in the figure has
a different depth. This means that when you view the figure in xfig, you
turn different dpeths on and off, emphasizing different information.

Depths make the diagram more useful, but they also make it much more difficult
to maintain, especially when adding new items or using xfig's update function.
To add new items, it is best to copy an existing item.  That way the new item
gets the depth of the existing item.


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@4931 908b024a-c82b-635c-c293-f2b15b23277c