BUGZID: 597 Reviewed by: Not reviewed
Apply javascript and form updates to allow multiple expression to be updated simultaneously.

Use checkboxes to select expression records for updating.

Each checkbox when clicked, updates the structure display. Anatomy items are highlighted if their stage range overlaps with the selected structure(s). If the item is expressed in all selected expression, the item's add button is checked.


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@5953 908b024a-c82b-635c-c293-f2b15b23277c