BUGZID: 496 Reviewed by: Not reviewed foo: copy ""  ;;; foo points to an empty string! foo: copy []  ;;; foo points to an empty block!
a string! is a series! of char!.
a block!  is a series! of anythings.

both being series! they share properties and functions that work on  series!

my guess now is you ended up with a block! that contained both a
string! and another block! which contained a bare word.

goob: [{Catalysis of the ... at Xaa-|-Gly bonds.} [EC:3.4.24.3] ]
replace/all goob "|" "\|"

does nothing because the first element in the block! is the string!
{Catalysis of the ...} which does not match "|"
the second element in the block! is the block! [EC:3.4.24.3] which also
does not match "|"  and it is done.


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@6315 908b024a-c82b-635c-c293-f2b15b23277c