BUGZID: 111 Reviewed by: to be reviewed bu Christian
Although this is not really the inital load of these files it may as well be.
We first recorded considering Ensembl links three years ago, however,
each of the several times we tried, they were deemed unworthy.

Now both Sanger and our curators have agreed that the subset of 1:1
links are 'not too bad' so this set of files allows ZFIN to be populated
with links to the current Ensembl release  where there is no chance
of competing gene models being being linked to.

when re-run the script will drop obsolete links and add new links,
unchanged links will remain unchanged.


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@7902 908b024a-c82b-635c-c293-f2b15b23277c