added a javascript function to send updates to patoupdate_addmutant.apg via xmlhttprequest.  The 'FX' table now has a far right column with these 'push to pato' links.
The change that may or may not stay a part of the page is that when the FX
table is getting created, I'm doing an SQL query for each row to see if
the entry is already in pato.  This is slow.  I did some timing tests, and
it looks like it adds 20% to the page loads.  The curators didn't seem to
mind so much, so it's going in, and we'll see if it's an issue in production.

Since the feature is clearly pretty important to them, it's probably worth
the time to redo those individual sql statements as a 'where exists..' part
of the FX table's big SQL statement.


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@9253 908b024a-c82b-635c-c293-f2b15b23277c