BUGZID: none Reviewed by: Not reviewed Date: Mon, 28 Jan 2008 10:24:54 -0800 From: Chris Mungall <cjm@fruitfly.org> To: Sierra Moxon <staylor@cs.uoregon.edu> Subject: Re: ZFIN ids vs OBO ids
Yes this is the kind of thing.

If this is to become a permanent report we should think in terms of a single
format used by all.. so I should have thought about this more before asking
you!

If it is a shared format then the ZDB ids should be prefixed with ZDB: -
though this isn't necessary for a ZFIN specific report

oh, I prefer tabs to pipes...


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@11959 908b024a-c82b-635c-c293-f2b15b23277c