BUGZID: 2188 Reviewed by: Not reviewed to prep for the meeting, I checked out the new snpd table that we made for jeff smith's data specifically.
It used to have a unique requirement on name, and on abbrev separately.  But, I think if we store both rs and ss numbers in here, we'll need to include the ss number in the unique constraint.

So, a combo of ss_number, name, and abbrev in each row must be unique now.
and, the ss_number must be unique in each row.

I also made the pub not null.  I'll add a default value to this table once we get the jeff smith pub created, etc...


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@12007 908b024a-c82b-635c-c293-f2b15b23277c