BUGZID: 2366 Reviewed by: Not reviewed there are 3 js functions that pass variables to do-markerupdate.apg when a curator adds references to a sequence.  1 of these functions also passes a fdbcont_id to do-markerupdate.apg.
In do-markerupdate.apg, no check was performed to see that the fdbcont_id
was passed.  So, the 2 js functions that were not passing fdbcont_ids forced
the "no rows returned" result from a check in dblink for the dblink_zdb_id
that matched the fdbcont, acc_num, and marker id sent in.

So, added a check to see that fdbcont_id exists, and execute query appropriately.


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@12038 908b024a-c82b-635c-c293-f2b15b23277c