BUGZID: 2385 Reviewed by: Not reviewed
To support having multiple marker relation types show up in the reno redundancy screen, I added an additional tag that shows all of the encoding relations for an accession value.  This uses one additional method in the MarkerService class.  Abbreviation logic is stored explicitly in this class, but may need to be moved somewhere else at a later time.

Because the name of the previous tag was misleading, I moved it to GeneEncodedRelationLinkTag.

The tag was only added to redundancy_hit_list.jsp.


git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@12150 908b024a-c82b-635c-c293-f2b15b23277c