BugzID: 2639
Added a LookupTable entry point that extends the Lookup entrypoint, mirroring the behavior of the anatomy search in both (esp. with respect to synonyms).  

This was changed in two places in the code, the xpatselect and fishselect (mutant search).  Some refactoring needed to be done, as well, in order to allow for building both.




git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@12822 908b024a-c82b-635c-c293-f2b15b23277c