bugzid: 3034
In the last update to the code for attributing dblinks there was
a check to see if the xtra_param was empty, but since it gets a
default value of 'alias' - that couldn't happen.  I replaced that
with a check to see if the contents of $xtra_param was a valid 
fdbcont_zdb_id, then used it if it was.  I threw in two other small
changes, I added a prefix to the comment at the top and replaced
a select * from pub to see if the pub was valid to a 
'select 'x' from pub'  ..  imagine how much faster it must be now!




git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@14481 908b024a-c82b-635c-c293-f2b15b23277c