bugzid: 14116 Always sent author notification email to curator
This change also introduces a new instance variable to control whether to send the email
to the 'real' recipients. It is off by default so that dev sites don't accidentally send
email to external users. This was previously done by checking the domain, but a new variable
should make it more testable.