bugzid: 11604 Remove xpatview.apg; use a redirect to xpatselect instead.
xpatview was mostly a glorified redirect xpatselect. It did two things: 1) it fetched the gene
symbol based on the ZDB-GENE id. That logic is now in xpatselect. 2) it called fxallfigures.apg if
a ZDB-XPAT id is provided. fxallfigures is long gone, and the logs don't show anyone using that
URL pattern (except for bots). So we don't really care about that use case anymore.

So I got rid of xpatview.apg and let apache handle the redirect.