PUB-232 Update author notification letter to include all data types
This changes the content of the author notification letter to be similar to the data summary
section of the publication page. To accommodate that, the main logic of generating those links
was moved out of JSP and into Java code so it could be shared between the pub view page and the
pub tracker.