ZFIN-6680 Add data list component
This adds a data list component and uses it for the citations list
as well as other places where a table with only one column was
previously used. This also decouples the data fetching/state
management from the presentation of the DataTable react component
so that the fetching and state part can be reused for the new
DataList component.