ZFIN-7501
Test is passing for creating the cascading associations when a morpholino is associated with a publication.  Test also for behavior with non-STRs--though this test is perhaps less robust.
Transaction handling is up a level from the service class so that the service class can be tested without committing transactions mid-test.  Add test to db suite.