ZFIN-7693
Image service test was failing due to temporary folder configured as absolute path.  Additionally, the test for creating thumbnails, etc. was passing when it should've been failing.  Reworked the test and corresponding code to pass that test.  Changed the path to imagemagick's convert from a hard-coded location (/bin/convert) to a configurable location.