ZFIN-7693
This test was failing due to an IndexOutOfBoundsException at accessions.get(0) in when clause.  However, the test seems that it may still need some attention.  I'm not sure why the assertion is so flexible in accepting either an empty set or dblinks?  I adjusted the input accession numbers to include some tests that the comment indicated would fail.