ZFIN-7816:
Mitigation that handles the issue of getting empty objects back from the API a bit more gracefully.
The rendered table omits the links where there should be more data from the API, but it still
shows as much of the table as possible (instead of failing with console error message).