ZFIN-7857 Add version number to shell script arguments and use PubMed journal abbr. (#163)
* ZFIN-7857:

The validateAllianceFiles.sh script is expecting the release version as the second command line arg, but none is provided.  This goes back to 11/4/21 AFAICT.

This adds that argument through the jenkins task and exits the shell script early if none is provided.

Use PubMed abbreviation for journals.

Clean up using intellij suggestions.