ZFIN-8035
When adding new column, populate it with default values.
In liquibase file, add some deduplication logic for rows that are identical except for a null value in protein_accession_number (dropping the null row).
Avoid adding duplicates when building the list of GafEntries in the gaf input parser. Favor the duplicate with the geneProductFormID.