ZFIN-7945 and ZFIN-7944 (#264)
* ZFIN 7945

Complete adding the functionality for associating genome location entries with publications.
Provide the same chromosomal location edit dialog for all NTRs
Add configuration options for a read-only view and an option to be blank when empty (instead of displaying "None")
Add a read-only view of "Chromosomal Locations". Show citation references on edit page.
Remove information that we don't display to end users.
Change formatting based on feedback
Add test for new functionality to associate a genomeLocation with a publication.