ZFIN-8293: Hotfix for Release-1136 v2 (#415)
* The former version of getABRegID had a "limit 1" clause.
That is no longer supported, so this change duplicates the same behavior.

Using a Text Block with pg_format style hql.

* ZFIN-8293: another hotfix to handle the empty results case for uniqueResult

This is a fix to the fix for ZFIN-8293. The first one handled a result set
with more than one row returned.  It didn't properly handle an empty result
set. This fixes that.