ZFIN-7002: Code Cleanup
There are a few places that generate links to a gbrowse interface.
The links are of the form "/gbrowse/Zv9/?name=ZDB-GENE-..." or "/action/gbrowse/?name=ENSDARG00..."
This commit adds the logic so that if jbrowse flag is enabled, the generated links are of the form "/action/jbrowse/byName?name=...".
The controller that handles that endpoint determines the parameters needed for a jBrowse link and it redirects to the appropriate
jBrowse interface.