Babel upgrade:
Address error we were getting when running npm build:
```
WARNING in ./react/components/ribbon/Ribbon.js 49:18-32
export 'default' (imported as 'style') was not found in './style.scss' (module has no exports)
 @ ./react/components/ribbon/index.js 1:0-30 4:0-58
 @ ./react/containers/PhenotypeRibbon.js
 @ ./react/containers/ lazy ^\.\/.*$ namespace object
 @ ./react/index.js 34:2-43
```

Apparently Ribbon.js was trying to import from style.scss. However, the css-loader part of webpack does not handle that case unless it is importing from a "module". Renaming the file to `style.module.scss` tells webpack that it is a module.