ZFIN-9112:
Remove buttons for "Format into a printable listing" that generate a printable list of citations for a given publication search due to expensive back-end solr query that is causing problems in solr.

Redirect to faceted search for "List all pubs" link.

Redirect to faceted search for any direct requests to the endpoints for "Format into a printable listing"