ZFIN-9282:
- Use environment variables instead of <!--|...|-->
- Fix join logic:
Previously, the sql logic for "Number of Genes with TALEN phenotype in a Tg background (TALEN used transiently)" was failing to include a join condition for joining fish_str and fish tables resulting in a cross join. I refactored it to explicitly use join operations to make the logic easier to read.